requirements.txt
langchain==0.1.13
streamlit==1.32.2
langchain-openai==0.1.1
beautifulsoup4==4.12.3
chromadb==0.4.24
python-dotenv==1.0.1
youtube-transcript-api==0.6.2